backend/requirements.txt
Flask
psycopg2-binary
PyJWT
flask-cors
web3==6.11.3
python-dotenv
py-solc-x==2.0.2
